requirements.txt
langchain==0.1.12
python-dotenv==1.0.0
streamlit==1.30.0
unstructured==0.9.2
tiktoken==0.7.0
faiss-cpu==1.8.0
libmagic==1.0
python-magic==0.4.27
google-generativeai==0.7.0
requests==2.32.2
